+++ spec/lib/addressfinder_spec.rb
@@ -77,7 +77,17 @@ RSpec.describe AddressFinder do
     let(:args){ {email: "john.doe@addressfinder.com"} }
 
     it "calls the email verification class" do
-      expect(AddressFinder::Email::Verification).to receive_message_chain(:new, :perform, :result)
+      expect(AddressFinder::V1::Email::Verification).to receive_message_chain(:new, :perform, :result)
+      subject
+    end
+  end
+
+  context "#phone_verification" do
+    subject(:verification) { AddressFinder.phone_verification(args) }
+    let(:args){ {phone_number: "1800 152 363", default_country_code: "AU"} }
+
+    it "calls the phone verification class" do
+      expect(AddressFinder::V1::Phone::Verification).to receive_message_chain(:new, :perform, :result)
       subject
     end
   end