+++ CHANGELOG.md
@@ -1,7 +1,9 @@
-# AddressFinder 1.9.2 (August 2023) #
+# AddressFinder 1.10.0 (August 2023) #
 
 * Add support for Email Verification
+* Add support for Phone Verification
 * Add support for Bulk Email Verification
+* Add support for Bulk Phone Verification
 
 # AddressFinder 1.9.1 (October 2022) #